% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{accumulate_bWeightHuman_Human}
\alias{accumulate_bWeightHuman_Human}
\title{Accumulate Patch Biting Weight}
\usage{
accumulate_bWeightHuman_Human()
}
\description{
Update my contribution to my current patch human biting weight by calling \code{\link{accumulate_bWeightHuman_MacroPatch}}
}
\details{
\itemize{
\item This method is bound to \code{Human$accumulate_bWeightHuman}
}
}
