% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{accumulate_bWeightZoo_MacroPatch}
\alias{accumulate_bWeightZoo_MacroPatch}
\title{MacroPatch: Accumulate Patch Zoo Biting Weight}
\usage{
accumulate_bWeightZoo_MacroPatch(bWeightZoo)
}
\arguments{
\item{bWeightZoo}{numeric}
}
\description{
Accumulate \code{private$bWeightZoo} (equivalent to private$bWeightZoo += bWeightZoo)
}
