% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{accumulate_bWeightZootox_MacroPatch}
\alias{accumulate_bWeightZootox_MacroPatch}
\title{MacroPatch: Accumulate Patch Zootox Biting Weight}
\usage{
accumulate_bWeightZootox_MacroPatch(bWeightZootox)
}
\arguments{
\item{bWeightZootox}{numeric}
}
\description{
Accumulate \code{private$bWeightZootox} (equivalent to private$bWeightZootox += bWeightZootox)
}
