% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{accumulate_kappa_MacroPatch}
\alias{accumulate_kappa_MacroPatch}
\title{MacroPatch: Accumulate Patch \eqn{\Kappa}}
\usage{
accumulate_kappa_MacroPatch(kappa)
}
\arguments{
\item{kappa}{numeric}
}
\description{
\eqn{\Kappa} is the proportion of mosquitoes that would become infected after blood feeding on any human, the net infectiousness of humans.
Accumulate \code{private$kappa} (equivalent to private$kappa += kappa)
}
