% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{add2Q_feverPfSI}
\alias{add2Q_feverPfSI}
\title{PfSI \code{Human} Event: Add PfSI Fever Event to Event Queue}
\usage{
add2Q_feverPfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of event}

\item{PAR}{\code{NULL}}
}
\description{
Add PfSI fever event to the event queue.
This method is called from \code{\link{infectHumanPfSI}}
This method adds event \code{\link{event_feverPfSI}} to the event queue.
This method is bound to \code{Human$add2Q_feverPfSI()}
}
