% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{add2Q_infectHumanPfSI}
\alias{add2Q_infectHumanPfSI}
\title{PfSI \code{Human} Event: Add PfSI Infection Event to Event Queue}
\usage{
add2Q_infectHumanPfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of infection}

\item{PAR}{\code{NULL}}
}
\description{
Add a PfSI infection to the event queue.
This method is called from \code{\link{infectiousBite_PfSI}}
This method adds event \code{\link{event_infectHumanPfSI}} to the event queue.
This method is bound to \code{Human$add2Q_infectHumanPfSI()}
}
