% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{add2Q_pewanePfMOI}
\alias{add2Q_pewanePfMOI}
\title{PfMOI \code{Human} Event: Add PfMOI PE Waning Protection Event to Event Queue}
\usage{
add2Q_pewanePfMOI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of vaccination}

\item{PAR}{\code{NULL}}
}
\description{
Add PfMOI PE waning protection event to the event queue.
This method is called from \code{\link{pevaccinatePfMOI}}
This method adds event \code{\link{event_pewanePfMOI}} to the event queue.
This method is bound to \code{Human$add2Q_pewanePfMOI()}
}
