% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Move.R
\name{add2Q_returnHome}
\alias{add2Q_returnHome}
\title{Move \code{Human} Event: Add Return Home Trip to Event Queue}
\usage{
add2Q_returnHome(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time to return home}

\item{PAR}{\code{NULL}}
}
\description{
Add a trip back home to my event queue.
This method is called from \code{\link{takeTrip}}
This method adds event \code{\link{event_returnHome}} to the event queue.
\itemize{
\item This method is bound to \code{Human$add2Q_returnHome()}
}
}
