% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-SimBite.R
\name{add2Q_SimBitePfSI_Human}
\alias{add2Q_SimBitePfSI_Human}
\title{PfSI SimBite \code{Human} Event: Add PfSI Simulated Bite Event to Event Queue}
\usage{
add2Q_SimBitePfSI_Human(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of bite}

\item{PAR}{do not set to a value (default \code{NULL})}
}
\description{
Add a simulated bite to a human at time given by \code{tEvent}. A \code{\link[MASHcpp]{mosquitoPfSI}} object
is created for the duration of this event.
}
\details{
\itemize{
\item This method is bound to \code{Human$add2Q_SimBitePfSI}
}
}
