% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Move.R
\name{add2Q_takeTrip}
\alias{add2Q_takeTrip}
\title{Move \code{Human} Event: Add a Trip to Event Queue}
\usage{
add2Q_takeTrip(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of trip}

\item{PAR}{\code{NULL}}
}
\description{
Add a trip to another patch to my event queue.
This method is called from \code{\link{initialize_trip_Human}}
This method adds event \code{\link{event_takeTrip}} to the event queue.
\itemize{
\item This method is bound to \code{Human$add2Q_takeTrip()}
}
}
