% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-ELP-Setup.R
\name{AQUA.ELP.MACRO.SETUP}
\alias{AQUA.ELP.MACRO.SETUP}
\title{Aquatic Ecology: ELP Setup}
\usage{
AQUA.ELP.MACRO.SETUP(overwrite = TRUE)
}
\arguments{
\item{overwrite}{overwrite existing methods}
}
\description{
Initialize methods.
}
\section{Classes}{

\itemize{
\item MacroPatch: \code{\link{get_ELP_MacroPatch}} bound to \code{MacroPatch$get_ELP}
\item MacroPatch: \code{\link{set_ELP_MacroPatch}} bound to \code{MacroPatch$set_ELP}
}
}

