% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-ELPool-Setup.R
\name{AQUA.ELPool.SETUP}
\alias{AQUA.ELPool.SETUP}
\title{Aquatic Ecology: ELPool Setup}
\usage{
AQUA.ELPool.SETUP(overwrite = TRUE)
}
\arguments{
\item{overwrite}{overwrite existing methods}
}
\description{
Initialize methods.
}
\section{Classes}{

\itemize{
\item MacroPatch: \code{\link{get_ELPool_MacroPatch}} bound to \code{MacroPatch$get_ELPool}
\item MacroPatch: \code{\link{set_ELPool_MacroPatch}} bound to \code{MacroPatch$set_ELPool}
}
}

