% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Base-Class.R
\docType{class}
\name{AquaPop_Base}
\alias{AquaPop_Base}
\title{Aquatic Population Abstract Base Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
AquaPop_Base
}
\description{
An abstract base class that specific aquatic ecology models will inherit from. This is not strictly necessary in the R6 object-oriented framework but it is intended
to mirror a C++ polymorphic implementation.
}
\section{\strong{Constructor}}{

\itemize{
\item The constructor (initialize) method must be overwritten by all inheriting classes.
}
}

\section{\strong{Methods}}{

\itemize{
\item oneDay_popDynamics: this method must be overwritten by all inheriting classes (pure virtual function in C++)
\item oneDay_addCohort: see \code{\link{oneDay_addCohort_AquaPop_Base}}
\item get_PatchPointer: see \code{\link{get_PatchPointer_AquaPop_Base}}
\item set_PatchPointer: see \code{\link{set_PatchPointer_AquaPop_Base}}
\item get_TilePointer: see \code{\link{get_TilePointer_AquaPop_Base}}
\item set_TilePointer: see \code{\link{set_TilePointer_AquaPop_Base}}
\item get_MosquitoPointer: see \code{\link{get_MosquitoPointer_AquaPop_Base}}
\item set_MosquitoPointer: see \code{\link{set_MosquitoPointer_AquaPop_Base}}
}
}

\section{\strong{Fields}}{

\itemize{
\item EggQ: set to \code{NULL} by default, not required by 'Emerge' model of aquatic ecology; see \code{\link[MASHcpp]{EggQ}}
\item ImagoQ: see \code{\link[MASHcpp]{ImagoQ}}
\item PatchPointer: a reference to a \code{\link{MacroPatch}} object
\item TilePointer: a reference to a \code{\link{MacroTile}} object
\item MosquitoPointer: a reference to an object inheriting from \code{\link{Mosquito_Base}}
}
}

\keyword{R6}
\keyword{class}
