% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Emerge-Parameters.R
\name{AquaPop_Emerge.Parameters}
\alias{AquaPop_Emerge.Parameters}
\title{Generate Parameters for Emerge Aquatic Ecology Model}
\usage{
AquaPop_Emerge.Parameters(nPatch, lambda, seasonality = TRUE, offset = NULL)
}
\arguments{
\item{nPatch}{integer number of patches}

\item{lambda}{seasonal mean emergence for each patch}

\item{seasonality}{boolean flag for seasonality}

\item{offset}{optional offset in peak emergence}
}
\description{
Make parameters for Emerge
}
