% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Emerge-Class.R
\docType{class}
\name{AquaPop_Emerge}
\alias{AquaPop_Emerge}
\title{Aquatic Population Emerge Base Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
AquaPop_Emerge
}
\description{
This class inherits form \code{\link{AquaPop_Base}}, this documentation only records
overridden or new methods & fields, please see the base class documentation for more details.
}
\section{\strong{Constructor}}{

\itemize{
\item lambda: numeric vector of mean adult female emergence of length 365
}
}

\section{\strong{Methods}}{

\itemize{
\item oneDay_popDynamics: see \code{\link{oneDay_popDynamics_AquaPop_Emerge}}
}
}

\section{\strong{Fields}}{

\itemize{
\item lambda: numeric vector of mean adult female emergence of length 365
}
}

\keyword{R6}
\keyword{class}
