% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMECOL-BiteEvents.R
\name{biteIntensity_SimEcol}
\alias{biteIntensity_SimEcol}
\title{Simulated Ecology: Biting Intensity Function}
\usage{
biteIntensity_SimEcol(t, hb, hs, trend = 0, wt = 0)
}
\arguments{
\item{t}{time (units of days)}

\item{hb}{baseline biting}

\item{hs}{amplitude of seasonal biting around baseline \code{hb}}

\item{trend}{modulate biting intensity over time}

\item{wt}{biting weight as function of time (is a vector calculated from \code{\link{biteWeight_SimEcol}})}
}
\description{
Evaluate instantaneous biting intensity at time t. Biting intensity is modulated by periodic
seasonality.
}
