% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMECOL-BiteEvents.R
\name{biteWeight_SimEcol}
\alias{biteWeight_SimEcol}
\title{Simulated Ecology: Age-dependent Biting Weight}
\usage{
biteWeight_SimEcol(age, aa = 1.7, bb = 4)
}
\arguments{
\item{age}{numeric value of age, given in days since birth}

\item{aa}{maximum value attained by the function}

\item{bb}{speed at which function attains its maximum (smaller values lead to faster attainment of maximum, larger values become linear, must be >0)}
}
\description{
Evaluate age-dependent human biting weight.
}
