% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-HumanPop-Parameters.R
\name{bitingWeight_HumanPop}
\alias{bitingWeight_HumanPop}
\title{Generate Biting Weight on Humans}
\usage{
bitingWeight_HumanPop(age, aa = 1.7, bb = 4)
}
\arguments{
\item{age}{age of human in days}

\item{aa}{maximum of function}

\item{bb}{scaling parameter (values closer to 0 lead to faster attainment of asymptotic biting rate)}
}
\description{
Generate a biting weight as a function of age. This function is called in \code{\link{sitePops_HumanPop}}
}
\examples{
bitingWeight_HumanPop(age = 1:(365*70))
}
