% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{close_conMove_HumanPop}
\alias{close_conMove_HumanPop}
\title{\code{HumanPop} Method: Close Pathogen Output Connection}
\usage{
close_conMove_HumanPop()
}
\description{
Closes the \code{\link[base]{connection}} object that is responsible for movement output
\itemize{
\item This method is bound to \code{HumanPop$close_conMove}
}
}
