% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{close_conPathogen_HumanPop}
\alias{close_conPathogen_HumanPop}
\title{\code{HumanPop} Method: Close Pathogen Output Connection}
\usage{
close_conPathogen_HumanPop()
}
\description{
Closes the \code{\link[base]{connection}} object that is responsible for pathogen model output
\itemize{
\item This method is bound to \code{HumanPop$close_conPathogen}
}
}
