% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Simulation.R
\name{closeCon_Tile}
\alias{closeCon_Tile}
\title{Close Output Connections}
\usage{
closeCon_Tile()
}
\description{
Close \code{\link[base]{connection}} objects
}
\details{
\itemize{
\item This method is bound to \code{MacroTile$closeCon}
}
}
