% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEBUG.R
\name{DEBUG.MASHMACRO}
\alias{DEBUG.MASHMACRO}
\title{MASH-MACRO: Debug Methods for MACRO Objects}
\usage{
DEBUG.MASHMACRO(overwrite = TRUE)
}
\description{
Print when objects are garbage collected and other additional methods. For debugging purposes only.
}
\details{
\itemize{
\item Human:
\itemize{
\item finalize: print myID when garbage collected
}
}
}
