% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{decrement_bWeightHuman_Human}
\alias{decrement_bWeightHuman_Human}
\title{Decrement Patch Biting Weight}
\usage{
decrement_bWeightHuman_Human()
}
\description{
Update my contribution to my current patch human biting weight by calling \code{\link{decrement_bWeightHuman_MacroPatch}}
}
\details{
\itemize{
\item This method is bound to \code{Human$decrement_bWeightHuman}
}
}
