% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{decrement_bWeightHuman_MacroPatch}
\alias{decrement_bWeightHuman_MacroPatch}
\title{MacroPatch: Decrement Patch Human Biting Weight}
\usage{
decrement_bWeightHuman_MacroPatch(bWeightHuman)
}
\arguments{
\item{bWeightHuman}{numeric}
}
\description{
Decrement \code{private$bWeightHuman} (equivalent to private$bWeightHuman -= bWeightHuman)
}
