% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{event_endPfSI}
\alias{event_endPfSI}
\title{PfSI \code{Human} Event: Generate PfSI Clearance Event}
\usage{
event_endPfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of clearance}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfSI clearance event to place in event queue.
This method is called from \code{\link{add2Q_endPfSI}}
This method is bound to \code{Human$event_endPfSI()}
\itemize{
\item tag: \code{\link{endPfSI}}
\item tEvent: clearance time is calculated as tEnd = tEvent + \code{\link{PfSI_ttClearPf}}
}
}
