% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{event_feverPfSI}
\alias{event_feverPfSI}
\title{PfSI \code{Human} Event: Generate PfSI Fever Event}
\usage{
event_feverPfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of clearance}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfSI dever event to place in event queue.
This method is called from \code{\link{add2Q_feverPfSI}}
This method is bound to \code{Human$event_feverPfSI()}
\itemize{
\item tag: \code{\link{feverPfSI}}
\item tEvent: fever time is calculated as tFever = tEvent + \code{\link{PfSI_ttFeverPf}}
}
}
