% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{event_gsvaccinatePfSI}
\alias{event_gsvaccinatePfSI}
\title{PfSI \code{Human} Event: Generate PfSI GS Vaccination Event}
\usage{
event_gsvaccinatePfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{begin gs protection}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfSI GS vaccination event to place in event queue.
This method is called from \code{\link{add2Q_gsvaccinatePfSI}}
This method is bound to \code{Human$event_gsvaccinatePfSI()}
\itemize{
\item tag: \code{\link{gsvaccinatePfSI}}
}
}
