% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{event_gswanePfSI}
\alias{event_gswanePfSI}
\title{PfSI \code{Human} Event: Generate PfSI GS Waning Protection Event}
\usage{
event_gswanePfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{end gs protection}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfSI GS waning protection event to place in event queue.
This method is called from \code{\link{add2Q_gsvaccinatePfSI}}
This method is bound to \code{Human$event_gsvaccinatePfSI()}
\itemize{
\item tag: \code{\link{gsvaccinatePfSI}}
\item tEvent: loss of efficacy is calculated as tWane = tEvent + \code{\link{PfSI_ttGSWanePf}}
}
}
