% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{event_pewanePfSI}
\alias{event_pewanePfSI}
\title{PfSI \code{Human} Event: Generate PfSI PE Waning Protection Event}
\usage{
event_pewanePfSI(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{end PE protection}

\item{PAR}{\code{NULL}}
}
\description{
Generate PfSI PE waning protection event to place in event queue.
This method is called from \code{\link{add2Q_pevaccinatePfSI}}
This method is bound to \code{Human$event_pevaccinatePfSI()}
\itemize{
\item tag: \code{\link{pevaccinatePfSI}}
\item tEvent: loss of efficacy is calculated as tWane = tEvent + \code{\link{PfSI_ttPEWanePf}}
}
}
