% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Move.R
\name{event_returnHome}
\alias{event_returnHome}
\title{Move \code{Human} Event: Generate a Return Home Event}
\usage{
event_returnHome(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time to return home}

\item{PAR}{\code{NULL}}
}
\description{
Generate a return home event to place in event queue.
This method is called from \code{\link{add2Q_returnHome}}
This method is bound to \code{Human$event_returnHome()}
\itemize{
\item tag: \code{\link{returnHome}}
}
}
