% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-SimBite.R
\name{event_SimBitePfSI_Human}
\alias{event_SimBitePfSI_Human}
\title{PfSI SimBite \code{Human} Event: Generate PfSI Simulated Bite Event}
\usage{
event_SimBitePfSI_Human(tEvent, PAR = NULL)
}
\arguments{
\item{tEvent}{time of bite}

\item{PAR}{do not set to a value (default \code{NULL})}
}
\description{
Generate a simulated bite event to place in the \code{\link[MASHcpp]{EventQ}};
this method should only be called through \code{\link{add2Q_SimBitePfSI_Human}}
}
\details{
\itemize{
\item This method is bound to \code{Human$event_SimBitePfSI}
}
}
