% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Move.R
\name{event_takeTrip}
\alias{event_takeTrip}
\title{Move \code{Human} Event: Generate a Trip Event}
\usage{
event_takeTrip(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of trip}

\item{PAR}{\code{NULL}}
}
\description{
Generate a trip event to place in event queue.
This method is called from \code{\link{add2Q_takeTrip}}
This method is bound to \code{Human$event_takeTrip()}
\itemize{
\item tag: \code{\link{takeTrip}}
}
}
