% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{feverPfSI}
\alias{feverPfSI}
\title{PfSI \code{Human} Event: PfSI Fever Event}
\usage{
feverPfSI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of fever}

\item{PAR}{\code{NULL}}
}
\description{
Simulate a PfSI fever event.
This method is bound to \code{Human$feverPfSI()}
\itemize{
\item \code{\link{add2Q_treatPfSI}}: fever may initiate treatment with probability \code{TreatPf}, see \code{\link{PfSI.Parameters}}
}
}
