% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{get_bWeightHuman_MacroPatch}
\alias{get_bWeightHuman_MacroPatch}
\title{MacroPatch: Get Patch Human Biting Weight}
\usage{
get_bWeightHuman_MacroPatch()
}
\description{
Return \code{private$bWeightHuman}
}
