% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{get_bWeightZoo_MacroPatch}
\alias{get_bWeightZoo_MacroPatch}
\title{MacroPatch: Get Patch Zoo Biting Weight}
\usage{
get_bWeightZoo_MacroPatch()
}
\description{
Return \code{private$bWeightZoo}
}
