% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{get_bWeightZootox_MacroPatch}
\alias{get_bWeightZootox_MacroPatch}
\title{MacroPatch: Get Patch Zootox Biting Weight}
\usage{
get_bWeightZootox_MacroPatch()
}
\description{
Return \code{private$bWeightZootox}
}
