% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{get_conPathogen_HumanPop}
\alias{get_conPathogen_HumanPop}
\title{\code{HumanPop} Method: Get Pathogen Output Connection}
\usage{
get_conPathogen_HumanPop()
}
\description{
Gets the \code{\link[base]{connection}} object that is responsible for pathogen model output
\itemize{
\item This method is bound to \code{HumanPop$get_conPathogen}
}
}
