% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Base-Class.R
\name{get_EggQ_AquaPop_Base}
\alias{get_EggQ_AquaPop_Base}
\title{Get Egg Queue}
\usage{
get_EggQ_AquaPop_Base()
}
\description{
Returns a reference to the \code{\link[MASHcpp]{EggQ}}
}
\details{
\itemize{
\item This method is bound to \code{AquaPop_Base$get_EggQ}
}
}
