% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{get_EIR_Human}
\alias{get_EIR_Human}
\title{Get \code{Human} EIR}
\usage{
get_EIR_Human()
}
\description{
Return my personal EIR value.
}
\details{
\itemize{
\item This method is bound to \code{Human$get_EIR}
}
}
