% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-ELP-Methods.R
\name{get_ELP_MacroPatch}
\alias{get_ELP_MacroPatch}
\title{Aquatic Eceology ELP: Getter for \code{\link{ELP}} in \code{\link{MacroPatch}}}
\usage{
get_ELP_MacroPatch()
}
\description{
Return \code{\link{ELP}}
\itemize{
\item This method is bound to \code{MacroPatch$get_ELP}
}
}
