% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Simulation.R
\name{get_emergingAdults_Mosquito_RM}
\alias{get_emergingAdults_Mosquito_RM}
\title{Recieve Emerging Adults}
\usage{
get_emergingAdults_Mosquito_RM(M, ix)
}
\description{
Recieves a number of emerging adults in a certain patch. This method directly interfaces with \code{\link{oneDay_addCohort_AquaPop_Base}}
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_RM$get_emergingAdults}
}
}
