% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{get_history_HumanPop}
\alias{get_history_HumanPop}
\title{HumanPop: Get all Histories}
\usage{
get_history_HumanPop()
}
\description{
Return all human histories as a named list (names correspond to human \code{myID}). See \code{\link[MASHcpp]{HistoryGeneric}} for details on the history implementation.
}
