% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{get_human_HumanPop}
\alias{get_human_HumanPop}
\title{HumanPop: Get a Human}
\usage{
get_human_HumanPop(humanID)
}
\arguments{
\item{humanID}{character}
}
\description{
Given a character \code{humanID} return that human (calls \code{\link[MASHcpp]{get_HashMap}}) to get the human.
}
