% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Class.R
\name{get_HumansPointer_MacroTile}
\alias{get_HumansPointer_MacroTile}
\title{Get Patches}
\usage{
get_HumansPointer_MacroTile()
}
\description{
Return a reference to the \code{\link{HumanPop}} residing in this tile
}
\details{
\itemize{
\item This method is bound to \code{MacroTile$get_HumansPointer}
}
}
