% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Base-Class.R
\name{get_HumansPointer_Mosquito_Base}
\alias{get_HumansPointer_Mosquito_Base}
\title{Get Humans Pointer}
\usage{
get_HumansPointer_Mosquito_Base()
}
\description{
Return a pointer to this \code{\link{HumanPop}} in this tile.
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_Base$get_HumansPointer}
}
}
