% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Base-Class.R
\name{get_ImagoQ_AquaPop_Base}
\alias{get_ImagoQ_AquaPop_Base}
\title{Get Imago Queue}
\usage{
get_ImagoQ_AquaPop_Base()
}
\description{
Returns a reference to the \code{\link[MASHcpp]{ImagoQ}}
}
\details{
\itemize{
\item This method is bound to \code{AquaPop_Base$get_ImagoQ}
}
}
