% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{get_kappa_MacroPatch}
\alias{get_kappa_MacroPatch}
\title{MacroPatch: Get Patch \eqn{\Kappa}}
\usage{
get_kappa_MacroPatch()
}
\description{
\eqn{\Kappa} is the proportion of mosquitoes that would become infected after blood feeding on any human, the net infectiousness of humans.
Return \code{private$kappa}
}
