% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Class.R
\name{get_M_Mosquito_RM}
\alias{get_M_Mosquito_RM}
\title{Get Mosquito Density for all Patches}
\usage{
get_M_Mosquito_RM(patch = NULL)
}
\description{
Return the vector of mosquito densities for all patches
}
\details{
\itemize{
\item This method is  bound to \code{Mosquito_RM$get_M}
}
}
