% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Base-Class.R
\name{get_MosquitoPointer_AquaPop_Base}
\alias{get_MosquitoPointer_AquaPop_Base}
\title{Get Mosquito Pointer}
\usage{
get_MosquitoPointer_AquaPop_Base()
}
\description{
Return a pointer to the mosquito population inheriting from \code{\link{Mosquito_Base}} in this tile.
}
