% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Class.R
\name{get_MosquitoPointer_MacroTile}
\alias{get_MosquitoPointer_MacroTile}
\title{Get Patches}
\usage{
get_MosquitoPointer_MacroTile()
}
\description{
Return a reference to a mosquito population derived from \code{\link{Mosquito_Base}} residing in this tile
}
\details{
\itemize{
\item This method is bound to \code{MacroTile$get_MosquitoPointer}
}
}
