% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Class.R
\name{get_nPatch_MacroTile}
\alias{get_nPatch_MacroTile}
\title{Get Number of Patches}
\usage{
get_nPatch_MacroTile()
}
\description{
write me
}
\details{
\itemize{
\item This method is bound to \code{MacroTile$get_nPatch}
}
}
