% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Class.R
\name{get_parameter_MosquitoRM}
\alias{get_parameter_MosquitoRM}
\title{MosquitoRM: Return a Named Parameter}
\usage{
get_parameter_MosquitoRM(key)
}
\arguments{
\item{key}{a character key; return that value from the named parameter list.}
}
\description{
Return a value from \code{private$parameters} given a character key.
}
