% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Class.R
\name{get_Patch_MacroTile}
\alias{get_Patch_MacroTile}
\title{Get a Patch}
\usage{
get_Patch_MacroTile(patchID)
}
\arguments{
\item{patchID}{a id of the patch that will be coerced to a character string key}
}
\description{
Return a reference to a \code{\link{MacroPatch}}
}
\details{
\itemize{
\item This method is bound to \code{MacroTile$get_Patch}
}
}
