% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Class.R
\name{get_Patches_MacroTile}
\alias{get_Patches_MacroTile}
\title{Get Patches}
\usage{
get_Patches_MacroTile()
}
\description{
Return a reference to the \code{\link[MASHcpp]{HashMap}} object that contains all patches
}
\details{
\itemize{
\item This method is bound to \code{MacroTile$get_Patches}
}
}
