% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\name{get_PathogensHistory_Human}
\alias{get_PathogensHistory_Human}
\title{Human: Set Pathogen History}
\usage{
get_PathogensHistory_Human()
}
\description{
Calls \code{Pathogens$get_history} which should be defined for all human-stage pathogen objects.
}
