% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{get_PathogensHistory_HumanPop}
\alias{get_PathogensHistory_HumanPop}
\title{HumanPop: Get all Pathogen histories}
\usage{
get_PathogensHistory_HumanPop()
}
\description{
Get all pathogen histories.
\itemize{
\item This function is bound to \code{HumanPop$get_PathogensHistory()}
}
}
