% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{get_reservoir_MacroPatch}
\alias{get_reservoir_MacroPatch}
\title{MacroPatch: Get whether Patch is a Reservoir, with fixed EIR}
\usage{
get_reservoir_MacroPatch()
}
\description{
Return \code{private$reservoir}
}
