% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Base-Class.R
\name{get_TilePointer_AquaPop_Base}
\alias{get_TilePointer_AquaPop_Base}
\title{Get Tile Pointer}
\usage{
get_TilePointer_AquaPop_Base()
}
\description{
Return a pointer to the enclosing \code{\link{MacroTile}}
}
