% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{get_TilePointer_HumanPop}
\alias{get_TilePointer_HumanPop}
\title{Get Tile Pointer}
\usage{
get_TilePointer_HumanPop()
}
\description{
Return either microsimulation \code{\link[MASHmicro]{Tile}} or macrosimulation \code{\link{Tile}} enclosing this site.
\itemize{
\item This method is bound to \code{HumanPop$get_TilePointer}
}
}
