% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{get_TilePointer_MacroPatch}
\alias{get_TilePointer_MacroPatch}
\title{MacroPatch: Get Patch Pointer to Enclosing \code{\link{MacroTile}}}
\usage{
get_TilePointer_MacroPatch()
}
\description{
Return \code{private$TilePointer}
}
