% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Base-Class.R
\name{get_TilePointer_Mosquito_Base}
\alias{get_TilePointer_Mosquito_Base}
\title{Get Tile Pointer}
\usage{
get_TilePointer_Mosquito_Base()
}
\description{
Return a pointer to the enclosing \code{\link{MacroTile}}
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_Base$get_TilePointer}
}
}
