% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{get_travelWeight_MacroPatch}
\alias{get_travelWeight_MacroPatch}
\title{MacroPatch: Get Patch Human Travel Weight}
\usage{
get_travelWeight_MacroPatch()
}
\description{
Return \code{private$get_travelWeight}
}
