% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Class.R
\name{get_Y_Mosquito_RM}
\alias{get_Y_Mosquito_RM}
\title{Get Incubating Mosquito Density for all Patches}
\usage{
get_Y_Mosquito_RM()
}
\description{
Return the vector of incubating mosquito densities for all patches
}
\details{
\itemize{
\item This method is  bound to \code{Mosquito_RM$get_Y}
}
}
