% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Class.R
\name{get_Z_Mosquito_RM}
\alias{get_Z_Mosquito_RM}
\title{Get Infectious Mosquito Density for all Patches}
\usage{
get_Z_Mosquito_RM()
}
\description{
Return the vector of infectious mosquito densities for all patches
}
\details{
\itemize{
\item This method is  bound to \code{Mosquito_RM$get_Y}
}
}
