% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Class.R
\name{get_ZZ_Mosquito_RM}
\alias{get_ZZ_Mosquito_RM}
\title{Get EIP Progressing Mosquito Density for all Patches}
\usage{
get_ZZ_Mosquito_RM()
}
\description{
Return the matrix of mosquitoes advancing through the EIP for all patches
}
\details{
\itemize{
\item This method is  bound to \code{Mosquito_RM$get_ZZ}
}
}
