% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{gswanePfSI}
\alias{gswanePfSI}
\title{PfSI \code{Human} Event: PfSI GS Waning Protection Event}
\usage{
gswanePfSI(tEvent, PAR)
}
\arguments{
\item{tEvent}{end GS protection}

\item{PAR}{\code{NULL}}
}
\description{
End PfSI GS protection.
This method is bound to \code{Human$gswanePfSI()}
\itemize{
\item protection: infected human to mosquito transmission efficiency is set back to \code{Pf_c}, see \code{\link{PfSI.Parameters}}
}
}
