% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\docType{class}
\name{Human}
\alias{Human}
\title{Human Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Human
}
\description{
This is a generic human being blah blah ...
Each instance of a \code{Human} lives in a \code{\link{HumanPop}}
}
\section{Constructor}{

\itemize{
\item item 1:
}
}

\section{Methods}{

\itemize{
\item get_myID: see \code{\link{get_myID_human}}
\item set_myID: see \code{\link{set_myID_human}}
\item get_houseID: see \code{\link{get_houseID_human}}
\item set_houseID: see \code{\link{set_houseID_human}}
\item get_patchID: see \code{\link{get_patchID_human}}
\item set_patchID: see \code{\link{et_patchID_human}}
\item get_bDay: see \code{\link{get_bDay_human}}
\item set_bDay: see \code{\link{set_bDay_human}}
\item get_EventQ: see \code{\link{get_EventQ_human}}
\item get_Alive: see \code{\link{get_Alive_human}}
\item set_Alive: see \code{\link{set_Alive_human}}
\item get_sex: see \code{\link{get_sex_human}}
\item set_sex: see \code{\link{set_sex_human}}
\item get_bWeight: see \code{\link{get_bWeight_human}}
\item set_bWeight: see \code{\link{set_bWeight_human}}
\item set_Pathogens: see \code{\link{set_Pathogens_human}}
\item get_HumansPointer: see \code{\link{get_HumansPointer_human}}
\item set_HumansPointer: see \code{\link{set_HumansPointer_human}}
\item oneEvent: see \code{\link{oneEvent_human}}
\item runEvent: see \code{\link{runEvent_human}}
\item liveLife: see \code{\link{liveLife_human}}
\item event_maxDeath: see \code{\link{event_maxDeath_human}}
\item death: see \code{\link{death_human}}
\item track_history: see \code{\link{track_history_human}}
\item get_history: see \code{\link{get_history_human}}
}
}

\section{Fields}{

\itemize{
\item \strong{myID}: character identifier of this human, first digits are integer ID and digits after underscore (_) are the patchID; houseID is not necessary because
location is not resolved to house level in MACRO and MICRO only needs houseID to distribute risk onto \code{\link{feedingSite}} rather than to store humans.
myID is needed to resolve location to patch level in order to move people between patches in either MICRO or MACRO.
}
}

\keyword{R6}
\keyword{class}
