% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\docType{class}
\name{HumanPop}
\alias{HumanPop}
\title{MASH-MACRO HumanPop Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
HumanPop
}
\description{
generate a single human population; they may live in a MacroPatch or a MicroPatch
Each instance of a \code{Human} lives in a \code{\link{HumanPop}}
}
\section{Constructor}{

\itemize{
\item item 1:
}
}

\section{Methods}{

\itemize{
\item \strong{get_Human}: retrieve human whose field 'myID' matches argument 'humanID' in keylist of pop field.
\item item 1:
\item item 1:
\item item 1:
\item item 1:
}
}

\keyword{R6}
\keyword{class}
