% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{HumanPop_get_PfMOI_PAR}
\alias{HumanPop_get_PfMOI_PAR}
\title{PfMOI \code{HumanPop} Method: Get PfMOI_PAR}
\usage{
HumanPop_get_PfMOI_PAR(ix)
}
\arguments{
\item{ix}{name of parameter to return \code{\link{PfMOI.Parameters}}}
}
\description{
Get named element of PfMOI_PAR. See \code{\link{PfMOI.Parameters}} for definitions.
\itemize{
\item This method is bound to \code{HumanPop$get_PfMOI_PAR()}
}
}
