% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{infectHumanPfMOI}
\alias{infectHumanPfMOI}
\title{PfMOI \code{Human} Event: PfMOI Infection Event}
\usage{
infectHumanPfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of infection}

\item{PAR}{must be a list containing character \code{PfID} and \code{vectorID}}
}
\description{
Simulate a PfMOI infection. If the human is not under chemoprophylaxis, begin queuing events for this clonal variant's infection process.
Specifically, a new PfID is drawn and a new infection is begun by calling \code{humanPfMOI$add_Infection}
A bernoulli event is drawn to determine if this infection produces fever; if so \code{\link{add2Q_feverPfMOI}} is called.
The end of this PfMOI infection is queued by \code{\link{add2Q_endPfMOI}}
}
\details{
\itemize{
\item This method is bound to \code{Human$infectHumanPfMOI()}
}
}
