% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{init_PfSI}
\alias{init_PfSI}
\title{PfSI \code{HumanPop} Method: Initialize PfSI Infections}
\usage{
init_PfSI(PfPR, tStart = 0)
}
\description{
Initialize PfSI infections with parasite prevalence PfPR in a human population.
This method is bound to \code{HumanPop$init_PfSI()}
}
