% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{init_PfSI_Human}
\alias{init_PfSI_Human}
\title{PfSI \code{Human} Method: Initialize PfSI Infection}
\usage{
init_PfSI_Human(PfPR)
}
\arguments{
\item{PfPR}{numeric; probability I am infected}
}
\description{
Initialize infection with given probability for a single Human. This must be called after \code{\link{Human_set_humanPfSI}} is used to initialize the pathogen object.
\itemize{
\item This method is bound to \code{Human$init_PfSI()}
}
}
