% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{init_PfSI_HumanPop}
\alias{init_PfSI_HumanPop}
\title{PfSI \code{HumanPop} Method: Initialize PfSI Infections}
\usage{
init_PfSI_HumanPop(PfPR)
}
\arguments{
\item{PfPR}{numeric; prevalence}
}
\description{
Initialize PfSI infections with parasite prevalence PfPR in a human population.
\itemize{
\item This method is bound to \code{HumanPop$init_PfSI()}
}
}
