% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Simulation.R
\name{initCon_Tile}
\alias{initCon_Tile}
\title{Initialize Output Connections}
\usage{
initCon_Tile()
}
\description{
Open \code{\link[base]{connection}} objects to write output
}
\details{
\itemize{
\item This method is bound to \code{MacroTile$initCon}
}
}
