% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{initialize_bWeightHuman_HumanPop}
\alias{initialize_bWeightHuman_HumanPop}
\title{Initialize Patch Biting Weights}
\usage{
initialize_bWeightHuman_HumanPop()
}
\description{
At beginning of a simulation, initialize all patch human biting weights.
}
\details{
\itemize{
\item This method is bound to \code{HumanPop$initialize_bWeightHuman}
}
}
