% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Simulation.R
\name{initialize_output_Mosquito_RM}
\alias{initialize_output_Mosquito_RM}
\title{Initialize Population Output}
\usage{
initialize_output_Mosquito_RM(con)
}
\description{
Initialize mosquito population output
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_RM$initialize_output}
}
}
