% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Move.R
\name{initialize_output_Move_HumanPop}
\alias{initialize_output_Move_HumanPop}
\title{Move \code{HumanPop} Method: Initialize Movement Output}
\usage{
initialize_output_Move_HumanPop()
}
\description{
Initialize output for movement model.
\itemize{
\item This method is bound to \code{HumanPop$initialize_output_Move()}
}
}
