% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{initialize_output_PfMOI_HumanPop}
\alias{initialize_output_PfMOI_HumanPop}
\title{PfMOI \code{HumanPop} Method: Initialize PfMOI Output}
\usage{
initialize_output_PfMOI_HumanPop()
}
\description{
Initialize output for PfMOI model.
\itemize{
\item This method is bound to \code{HumanPop$initialize_output_Pathogen()}
}
}
