% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{initialize_output_PfSI_HumanPop}
\alias{initialize_output_PfSI_HumanPop}
\title{PfSI \code{HumanPop} Method: Initialize PfSI Output}
\usage{
initialize_output_PfSI_HumanPop()
}
\description{
Initialize output for PfSI model.
\itemize{
\item This method is bound to \code{HumanPop$initialize_output_Pathogen()}
}
}
