% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{initialize_Pathogens_PfMOI_HumanPop}
\alias{initialize_Pathogens_PfMOI_HumanPop}
\title{PfMOI \code{HumanPop} Method: Initialize PfMOI Infections}
\usage{
initialize_Pathogens_PfMOI_HumanPop(PfPAR)
}
\arguments{
\item{PfPAR}{integer vector of starting MOI for each human}
}
\description{
Initialize PfMOI infections with given MOI for each human.
\itemize{
\item This method is bound to \code{HumanPop$initialize_Pathogens()}
}
}
