% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{initialize_Pathogens_PfSI_HumanPop}
\alias{initialize_Pathogens_PfSI_HumanPop}
\title{PfSI \code{HumanPop} Method: Initialize PfSI Infections}
\usage{
initialize_Pathogens_PfSI_HumanPop(PfPAR)
}
\arguments{
\item{PfPAR}{numeric vector of parasite prevalence in each patch}
}
\description{
Initialize PfSI infections with parasite prevalence PfPR in a human population.
\itemize{
\item This method is bound to \code{HumanPop$initialize_Pathogens()}
}
}
