% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{initialize_peVaxx_Human}
\alias{initialize_peVaxx_Human}
\title{PfSI \code{Human} Method: Initialize and Queue Vaccination Events for each Human}
\usage{
initialize_peVaxx_Human(vaxPar)
}
\description{
Initialize vaccinations for each human as defined in peVaxPar.
}
\details{
\itemize{
\item This method is bound to \code{Human$initialize_peVaxx}
}
}
