% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{initialize_peVaxx_HumanPop}
\alias{initialize_peVaxx_HumanPop}
\title{PfSI \code{HumanPop} Method: Initialize and Queue Vaccination Events}
\usage{
initialize_peVaxx_HumanPop(peVaxPar)
}
\description{
Initialize vaccinations for each human as defined in peVaxPar, where peVaxPar is a list indexed on each human's ID that describes the times when each human receives a PE vaccination.
}
\details{
\itemize{
\item This method is bound to \code{HumanPop$initialize_peVaxx}
}
}
