% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Move.R
\name{initialize_travel_Human}
\alias{initialize_travel_Human}
\title{Move \code{Human} Event: Initialize Travel}
\usage{
initialize_travel_Human()
}
\description{
Initialize travel model when simulation begins by queueing up a trip calling \code{\link{add2Q_takeTrip}}
\itemize{
\item This method is bound to \code{Human$initialize_travel}
}
}
