% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Move.R
\name{initialize_travel_HumanPop}
\alias{initialize_travel_HumanPop}
\title{Move \code{Human} Event: Initialize Travel}
\usage{
initialize_travel_HumanPop()
}
\description{
Initialize travel model when simulation begins by calling \code{\link{initialize_travel_Human}} for all humans.
\itemize{
\item This method is bound to \code{HumanPop$initialize_travel}
}
}
