% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Simulation.R
\name{initOutput_Mosquito_RM}
\alias{initOutput_Mosquito_RM}
\title{Initialize Population Output}
\usage{
initOutput_Mosquito_RM(con)
}
\description{
Initialize mosquito population output
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_RM$initOutput}
}
}
