% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-ELP-Fit.R
\name{K2psi_ELP}
\alias{K2psi_ELP}
\title{Aquatic Ecology: ELP Calculate Psi from K}
\usage{
K2psi_ELP(f, v, alpha, g, gamma, K)
}
\arguments{
\item{f}{adult female mosquito blood feeding rate (waiting time between bloodmeals is 1/f)}

\item{v}{size of egg batch per bloodmeal}

\item{alpha}{daily mautration rate of aquatic stages (1/alpha is expected duration of aquatic life stages)}

\item{g}{daily adult female mortality rate (average lifespan is 1/g)}

\item{gamma}{daily density independent mortality rate}

\item{K}{carrying capacity}
}
\value{
numeric
}
\description{
Calculate strength of density dependence parameter psi (increase in per-capita mortality in response to crowding) from
adult and pool characteristics. Functional form of equation given by \deqn{K=\sqrt[\sigma]{\frac{\frac{fv \alpha }{g}-(\alpha + \gamma)}{\psi }}}
For more details and references see \code{\link[MASHcpp]{ELP}}.
}
