% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{lmTest_PfSI}
\alias{lmTest_PfSI}
\title{PfSI \code{Human} Method: PfSI Light Microscopy Test}
\usage{
lmTest_PfSI(tEvent, PAR)
}
\description{
Administer light microscopy to this human.
\itemize{
\item if infected: true positive is detected with probability \code{lmSensPf}, see \code{\link{PfSI.Parameters}}
\item if susceptible: false positive is detected with probability \code{lmSpecPf}, see \code{\link{PfSI.Parameters}}
}
}
