% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Setup.R
\name{MACRO.Human.Setup}
\alias{MACRO.Human.Setup}
\title{MACRO Human Module Setup}
\usage{
MACRO.Human.Setup(pathogen = "PfSI")
}
\description{
Initialize movement and biting modules for MACRO Humans.
}
