% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Class.R
\docType{class}
\name{MacroTile}
\alias{MacroTile}
\title{MASH-MACRO MacroTile Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MacroTile
}
\description{
write me
}
\section{\strong{Constructor}}{

\itemize{
\item argument: i'm an argument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: i'm a field!
}
}

\keyword{R6}
\keyword{class}
