% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMECOL-BiteEvents.R
\name{makeBites_SimEcol}
\alias{makeBites_SimEcol}
\title{Simulated Ecology: Generate Seasonal Biting Pattern}
\usage{
makeBites_SimEcol(N, mxYR, hb, hs, trend = 0, dt = 5)
}
\arguments{
\item{N}{total number of bites over time interval}

\item{mxYR}{time in years to simulate (will run from 0 to mxYR)}

\item{hb}{passed to \code{\link{biteIntensity_SimEcol}}}

\item{hs}{passed to \code{\link{biteIntensity_SimEcol}}}

\item{trend}{strength of seasonal trend, passed to \code{\link{biteIntensity_SimEcol}}}

\item{dt}{time step (in days)}
}
\description{
Generate a seasonally-forced lifetime biting pattern for a single human as a non-homogeneous Poisson process.
}
