% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMECOL-BiteEvents.R
\name{makeBitesPlot_SimEcol}
\alias{makeBitesPlot_SimEcol}
\title{Simulated Ecology: Plot Seasonal Biting Pattern}
\usage{
makeBitesPlot_SimEcol(N, mxYR, hb, hs, trend = 0, dt = 5,
  returnVals = FALSE)
}
\arguments{
\item{N}{total number of bites over time interval}

\item{mxYR}{time in years to simulate (will run from 0 to mxYR)}

\item{hb}{passed to \code{\link{biteIntensity_SimEcol}}}

\item{hs}{passed to \code{\link{biteIntensity_SimEcol}}}

\item{trend}{strength of seasonal trend, passed to \code{\link{biteIntensity_SimEcol}}}

\item{dt}{time step (in days)}

\item{returnVals}{if \code{TRUE}, return the biting times}
}
\description{
Plot the non-homogeneous Poisson biting process produced by \code{\link{makeBitesPlot_SimEcol}}.
}
