% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Base-Class.R
\docType{class}
\name{Mosquito_Base}
\alias{Mosquito_Base}
\title{Mosquito Population Abstract Base Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Mosquito_Base
}
\description{
An abstract base class that specific mosquito ecology models will inherit from. This is not strictly necessary in the R6 object-oriented framework but it is intended
to mirror a C++ polymorphic implementation.
}
\section{\strong{Constructor}}{

\itemize{
\item The constructor (initialize) method must be overwritten by all inheriting classes.
}
}

\section{\strong{Methods}}{

\itemize{
\item oneDay_popDynamics: this method must be overwritten by all inheriting classes (pure virtual function in C++)
\item oneDay_oviposition: this method must be overwritten by all inheriting classes (pure virtual function in C++)
\item get_emergingAdults: this method must be overwritten by all inheriting classes (pure virtual function in C++)
\item get_HumansPointer: see \code{\link{get_HumansPointer_Mosquito_Base}}
\item set_HumansPointer: see \code{\link{set_HumansPointer_Mosquito_Base}}
\item get_TilePointer: see \code{\link{get_TilePointer_Mosquito_Base}}
\item set_TilePointer: see \code{\link{set_TilePointer_Mosquito_Base}}
}
}

\section{\strong{Fields}}{

\itemize{
\item HumansPointer: a reference to a \code{\link{HumanPop}} object
\item TilePointer: a reference to a \code{\link{MacroTile}} object
}
}

\keyword{R6}
\keyword{class}
