% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Class.R
\docType{class}
\name{Mosquito_RM}
\alias{Mosquito_RM}
\title{MosquitoRM Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Mosquito_RM
}
\description{
Generate a mosquito population living in a \code{\link{MacroTile}}, following Ross-MacDonald stype daily difference equations
and simple diffusion based movement given a transition matrix.
This class inherits form \code{\link{Mosquito_Base}}, this documentation only records
overridden or new methods & fields, please see the base class documentation for more details.
}
\section{\strong{Constructor}}{

\itemize{
\item M: numeric vector of adult female mosquito density of length equal to number of patches
\item EIP: numeric vector of daily EIP of length 365
\item p = 0.9: daily probability of survival
\item f = 0.3: daily feeding rate
\item Q = 0.9: human blood index
\item v = 20: egg batch size
\item psi = \code{NULL}: diffusion matrix, must have rows and columns equal to number of patches
}
}

\section{\strong{Methods}}{

\itemize{
\item oneDay_popDynamics: see \code{\link{oneDay_popDynamics_Mosquito_RM}}
\item oneDay_oviposition: see \code{\link{oneDay_oviposition_Mosquito_RM}}
\item get_emergingAdults: see \code{\link{get_emergingAdults_Mosquito_RM}}
}
}

\section{\strong{Fields}}{

\itemize{
\item i'm a field: write me
}
}

\keyword{R6}
\keyword{class}
