% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Class.R
\docType{class}
\name{MosquitoRM}
\alias{MosquitoRM}
\title{MosquitoRM Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MosquitoRM
}
\description{
Generate a single well-mixed mosquito population which lives in a \code{\link{MacroPatch}} which lives according to Ross-Macdonald style difference equations.
}
\section{\strong{Constructor}}{

}

\section{\strong{Methods}}{

\itemize{
\item get_pop: see \code{\link{get_pop_HumanPop}}
\item get_human: see \code{\link{get_human_HumanPop}}
\item get_history: see \code{\link{get_history_HumanPop}}
\item simHumans: see \code{\link{simHumans_HumanPop}}
}
}

\section{\strong{Fields}}{

\itemize{
\item N: number of human
\item tStart: time to start simulation
\item pop: a object of class \code{\link[MASHcpp]{HashMap}} that stores instantiations of \code{\link{Human}}, see help for more details on the internal structure of this type.
}
}

\keyword{R6}
\keyword{class}
