% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{normalize_kappa_MacroPatch}
\alias{normalize_kappa_MacroPatch}
\title{MacroPatch: Normalize Patch \eqn{\Kappa}}
\usage{
normalize_kappa_MacroPatch()
}
\description{
\eqn{\Kappa} is the proportion of mosquitoes that would become infected after blood feeding on any human, the net infectiousness of humans.
Normalize \eqn{\frac{\Kappa}{bWeightHuman + bWeightZoo + bWeightZootox}}
}
\details{
\itemize{
\item This method is bound to \code{MacroPatch$normalize_kappa}
}
}
