% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Base-Class.R
\name{oneDay_addCohort_AquaPop_Base}
\alias{oneDay_addCohort_AquaPop_Base}
\title{Add Emerging Cohort to Adult Population}
\usage{
oneDay_addCohort_AquaPop_Base(ix)
}
\arguments{
\item{ix}{the patch ID passed down from the calling \code{\link{MacroPatch}} in \code{\link{oneDay_addCohort_Patch}}}
}
\description{
Take emerging aquatic stage mosquitoes from \code{\link[MASHcpp]{ImagoQ}} and add to the adult population.
This function interacts with all mosquito models that inherit from \code{\link{Mosquito_Base}}, and requires
those models to have a \code{get_emergingAdults} method defined.
}
\details{
\itemize{
\item This method is bound to \code{AquaPop_Base$oneDay_addCohort}
}
}
