% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Simulation.R
\name{oneDay_addCohort_Patch}
\alias{oneDay_addCohort_Patch}
\title{Patch Aquatic Population Dynamics}
\usage{
oneDay_addCohort_Patch()
}
\description{
wrapper around the oneDay_addCohort of the embedded AquaPop_Base for polymorphism
}
