% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Emerge-Class.R
\name{oneDay_popDynamics_AquaPop_Emerge}
\alias{oneDay_popDynamics_AquaPop_Emerge}
\title{Run Emerge Daily Population Dynamics}
\usage{
oneDay_popDynamics_AquaPop_Emerge()
}
\description{
Calculate daily emerging adults with Poisson dispersion around mean value of lambda for this day and add them to the \code{\link[MASHcpp]{ImagoQ}}
}
\details{
\itemize{
\item This method is bound to \code{AquaPop_Emerge$oneDay_popDynamics}
}
}
