% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Simulation.R
\name{oneDay_popDynamics_Patch}
\alias{oneDay_popDynamics_Patch}
\title{Patch Aquatic Population Dynamics}
\usage{
oneDay_popDynamics_Patch()
}
\description{
wrapper around the oneDay_popDynamics of the embedded AquaPop_Base for polymorphism
}
