% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Simulation.R
\name{output_Mosquito_RM}
\alias{output_Mosquito_RM}
\title{Write Population Output}
\usage{
output_Mosquito_RM(con)
}
\description{
Write mosquito population output
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_RM$output}
}
}
