% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Simulation.R
\name{output_Patch}
\alias{output_Patch}
\title{Write Patch Output}
\usage{
output_Patch(con)
}
\description{
Write patch output
}
\details{
\itemize{
\item This method is bound to \code{MacroPatch$output}
}
}
