% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfLOME-Class.R
\docType{class}
\name{PfLOME_Human}
\alias{PfLOME_Human}
\title{PfLOME Human-stage Pathogen Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
PfLOME_Human
}
\description{
write me.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: write me
}
}

\section{\strong{Methods}}{

\itemize{
\item get_something: write me
\item set_something: write me
}
}

\section{\strong{Fields}}{

\itemize{
\item field: write me
}
}

\keyword{R6}
\keyword{class}
