% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{PfMOI_ttPEWanePf}
\alias{PfMOI_ttPEWanePf}
\title{PfMOI \code{Human} Method: Duration of protection by PE Vaccination}
\usage{
PfMOI_ttPEWanePf()
}
\description{
After administration of PE Vaccination what is duration of protection (reduced infected mosquito to human transmission efficiency)? See \code{\link{PfMOI.Parameters}} parameter \code{mnPEPf} and \code{vrPEPf} for normally distributed duration of protection.
This method is called from \code{\link{event_pewanePfMOI}}
This method is bound to \code{Human$ttPEWanePf()}
}
