% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{PfMOI_ttTreatPf}
\alias{PfMOI_ttTreatPf}
\title{PfMOI \code{Human} Method: Timing of Treatment}
\usage{
PfMOI_ttTreatPf()
}
\description{
What is the timing of treatment relative to the start of a fever incident? See \code{\link{PfMOI.Parameters}} parameter \code{mnTreatPf} for mean waiting time of exponentially distributed waiting period.
This method is called from \code{\link{}}
This method is bound to \code{Human$ttTreatPf()}
}
