% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Parameters.R
\name{PfSI.Parameters}
\alias{PfSI.Parameters}
\title{Initialize PfSI Module Parameters}
\usage{
PfSI.Parameters(Pf_c = 0.15, Pf_b = 0.55, DurationPf = 200,
  LatentPf = 10, FeverPf = 0.3, mnFeverPf = 10, vrFeverPf = 0.1,
  TreatPf = 0.5, mnTreatPf = 1, mnChemoprophylaxisPf = 32,
  PEProtectPf = 0.99, peBlockPf = 1, mnPEPf = 270, vrPEPf = 50,
  GSProtectPf = 1, gsBlockPf = 0.9, mnGSPf = 180, vrGSPf = 20,
  rdtSensPf = 0.9, rdtSpecPf = 0.1, lmSensPf = 0.9, lmSpecPf = 0.1)
}
\arguments{
\item{Pf_c}{0.15; infected human to mosquito transmission efficiency}

\item{Pf_b}{0.55; infected mosquito to human transmission efficiency}

\item{DurationPf}{200; duration of infection (How many days does the infection last?)}

\item{LatentPf}{10; latency (How many days after the infectious bite does the infection start?)}

\item{FeverPf}{0.3; probability of fever}

\item{mnFeverPf}{10; mean of Gaussian-distributed timing of fever incident (relative to start of the infection)}

\item{vrFeverPf}{0.1; standard deviation of Gaussian-distributed timing of fever incident (relative to start of the infection)}

\item{TreatPf}{0.5; probability of treatment}

\item{mnTreatPf}{1; timing of treatment (relative to start of the fever)}

\item{mnChemoprophylaxisPf}{32; Prophylaxis, time to susceptibility (duration of protection due to chemoprophylaxis)}

\item{PEProtectPf}{0.99; proportion protected by PE vaccination (probability vaccination successful)}

\item{peBlockPf}{1; proportion of infections blocked by PE vaccination}

\item{mnPEPf}{270; mean duration of protection from PE vaccination}

\item{vrPEPf}{50; standard deviation of protection from PE vaccination}

\item{GSProtectPf}{1; proportion protected by GS vaccination (probability vaccination successful)}

\item{gsBlockPf}{0.9; proportion of infections blocked by GS vaccination}

\item{mnGSPf}{180; mean duration of protection from GS vaccination}

\item{vrGSPf}{20; standard deviation of protection from GS vaccination}

\item{rdtSensPf}{0.9; RDT sensitivity}

\item{rdtSpecPf}{0.1; RDT specificity}

\item{lmSensPf}{0.9; Light Microscopy sensitivity}

\item{lmSpecPf}{0.1; Light Microscopy specificity}
}
\value{
return a list
}
\description{
This is mostly used to modify PfSI parameters for different human populations after creation by using \code{set_PfSI_PAR} method of HumanPop
}
\examples{
PfSI.Parameters()
}
