% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{PfSI_ttFeverPf}
\alias{PfSI_ttFeverPf}
\title{PfSI \code{Human} Method: Timing of Fever Incident}
\usage{
PfSI_ttFeverPf()
}
\description{
What is the timing of fever relative to the start of a PfSI infection? See \code{\link{PfSI.Parameters}} parameter \code{mnFeverPf} constant timing period.
This method is called from \code{\link{event_feverPfSI}}
This method is bound to \code{Human$ttFeverPf()}
}
