% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-SimBite.R
\name{queueBites_SimBitePfSI_Human}
\alias{queueBites_SimBitePfSI_Human}
\title{PfSI SimBite \code{Human} Event: Queue Bites}
\usage{
queueBites_SimBitePfSI_Human(tMax, bitingRate = 1/20)
}
\arguments{
\item{tMax}{maximum time to simulate}

\item{bitingRate}{biting rate}
}
\description{
Queue bites by simulating a homogeneous Poisson process and calling \code{\link{add2Q_SimBitePfSI_Human}}
}
\details{
\itemize{
\item This method is bound to \code{Human$queueBites_SimBitePfSI}
}
}
