% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{queueInfectiousBites_HumanPop}
\alias{queueInfectiousBites_HumanPop}
\title{Queue Daily Infectious Bites}
\usage{
queueInfectiousBites_HumanPop()
}
\description{
Based on personal EIR value, queue infectious bites for all humans.
}
\details{
\itemize{
\item This method is bound to \code{HumanPop$queueInfectiousBites}
}
}
