% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{queueInfectiousBites_PfMOI_Human}
\alias{queueInfectiousBites_PfMOI_Human}
\title{PfMOI Queue my Daily Infectious Bites}
\usage{
queueInfectiousBites_PfMOI_Human()
}
\description{
Based on my personal EIR value, queue infectious PfMOI bites by calling \code{\link{add2Q_SimBitePfMOI_Human}}
}
\details{
\itemize{
\item This method is bound to \code{Human$queueInfectiousBites}
}
}
