% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{queuePEVaccination_PfSI_Human}
\alias{queuePEVaccination_PfSI_Human}
\title{PfSI SimBite \code{Human} Event: Queue Vaccination}
\usage{
queuePEVaccination_PfSI_Human(tVaccine, tTreat = NA)
}
\arguments{
\item{tVaccine}{time of vaccination}

\item{tTreat}{time of treatment (may be \code{NULL})}
}
\description{
Queue vaccination and optionally treatment events by calling \code{\link{add2Q_pevaccinatePfSI}} and \code{\link{add2Q_treatPfSI}}
}
\details{
\itemize{
\item This method is bound to \code{Human$queuePEVaccination_PfSI}
}
}
