% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-SimBite.R
\name{queueVaccination_SimBitePfSI_Human}
\alias{queueVaccination_SimBitePfSI_Human}
\title{PfSI SimBite \code{Human} Event: Queue Vaccination}
\usage{
queueVaccination_SimBitePfSI_Human(tVaccine, tTreat = NULL)
}
\arguments{
\item{tVaccine}{time of vaccination}

\item{tTreat}{time of treatment (may be \code{NULL})}
}
\description{
Queue vaccination and optionally treatment events by calling \code{\link{add2Q_pevaccinatePfSI}} and \code{\link{add2Q_treatPfSI}}
}
\details{
\itemize{
\item This method is bound to \code{Human$queueVaccination_SimBitePfSI}
}
}
