% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Emerge-Class.R
\name{reset_AquaPop_Emerge}
\alias{reset_AquaPop_Emerge}
\title{Reset the Aquatic Population}
\usage{
reset_AquaPop_Emerge()
}
\description{
Reset this aquatic population between simulation runs
}
\details{
\itemize{
\item This method is bound to \code{AquaPop_Emerge$reset}
}
}
