% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\name{reset_Human}
\alias{reset_Human}
\title{\code{Human}: Reset Human Between Runs}
\usage{
reset_Human(houseID = NULL, patchID = NULL, homeHouseID = NULL,
  homePatchID = NULL, age = NULL, bWeight = NULL)
}
\arguments{
\item{houseID}{integer ID of my current house}

\item{patchID}{integer ID of my current patch}

\item{homeHouseID}{integer ID of my home house}

\item{homePatchID}{integer ID of my home patch}

\item{age}{numeric birthday (negative; simulation begins at 0)}

\item{bWeight}{numeric biting weight}

\item{myID}{a character ID}
}
\description{
Resets fields to default constructor values and clears out the \code{\link[MASHcpp]{EventQ}}
}
