% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{reset_HumanPop}
\alias{reset_HumanPop}
\title{\code{HumanPop}: Reset HumanPop Between Runs}
\usage{
reset_HumanPop(HumanPop_PAR)
}
\arguments{
\item{HumanPop_PAR}{list of length equal to number of humans}
}
\description{
Resets all humans by calling \code{\link{reset_Human}} for all humans; see that method for more details
}
