% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Simulation.R
\name{reset_Mosquito_RM}
\alias{reset_Mosquito_RM}
\title{Reset the Mosquito Population}
\usage{
reset_Mosquito_RM(MosquitoPar)
}
\description{
Reset mosquito population between simulation runs
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_RM$reset}
}
}
