% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Simulation.R
\name{reset_Patch}
\alias{reset_Patch}
\title{Reset Patch}
\usage{
reset_Patch(bWeightZoo, bWeightZootox)
}
\description{
Reset patch between simulation runs
}
\details{
\itemize{
\item This method is bound to \code{MacroPatch$reset}
}
}
