% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Simulation.R
\name{resetMacro}
\alias{resetMacro}
\title{Reset Tile between Runs}
\usage{
resetMacro(PatchPar, MosquitoPar)
}
\description{
write me
}
\details{
\itemize{
\item This method is bound to \code{MacroTile$resetMacro}
}
}
