% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Move.R
\name{returnHome}
\alias{returnHome}
\title{Move \code{Human} Event: Return Home Event}
\usage{
returnHome(tEvent, PAR)
}
\arguments{
\item{tEvent}{time to return home}

\item{PAR}{\code{NULL}}
}
\description{
Simulate my return home
This method is bound to \code{Human$returnHome()}
\itemize{
\item Biting weight at my origin is decremented by \code{\link{decrement_bWeightHuman_Human}} and incremented at my destination by \code{\link{accumulate_bWeightHuman_Human}}
\item My next trip is queued after a duration at home by calling \code{\link{add2Q_takeTrip}}
}
}
