% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Methods.R
\name{run_isolated_MacroPatch}
\alias{run_isolated_MacroPatch}
\title{MacroPatch: Run Macrosimulation for Isolated Patch}
\usage{
run_isolated_MacroPatch(tMax)
}
\arguments{
\item{tMax}{maximum time to run simulation}
}
\description{
Run a single \code{\link{MacroPatch}} as an isolated unit.
Keep in mind that human travel must be correctly initialized for this to work.
}
