% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoRM-Methods.R
\name{run_MigrationOut_MosquitoRM}
\alias{run_MigrationOut_MosquitoRM}
\alias{run_MigrationOut_MosquitoRM}
\title{MosquitoRM: Run Daily Inter-patch Outbound Migration}
\usage{
run_MigrationOut_MosquitoRM(M_in, Y_in, Z_in, Y0_in)

run_MigrationOut_MosquitoRM(M_in, Y_in, Z_in, Y0_in)
}
\arguments{
\item{some}{parameter}

\item{some}{parameter}
}
\description{
do something.

do something.
}
