% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{set_bWeightHuman_MacroPatch}
\alias{set_bWeightHuman_MacroPatch}
\title{MacroPatch: Set Patch Human Biting Weight}
\usage{
set_bWeightHuman_MacroPatch(bWeightHuman)
}
\arguments{
\item{bWeightHuman}{numeric}
}
\description{
Set \code{private$bWeightHuman}
}
