% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{set_bWeightZoo_MacroPatch}
\alias{set_bWeightZoo_MacroPatch}
\title{MacroPatch: Set Patch Zoo Biting Weight}
\usage{
set_bWeightZoo_MacroPatch(bWeightZoo)
}
\arguments{
\item{bWeightZoo}{numeric}
}
\description{
Set \code{private$bWeightZoo}
}
