% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{set_bWeightZootox_MacroPatch}
\alias{set_bWeightZootox_MacroPatch}
\title{MacroPatch: Set Patch Zootox Biting Weight}
\usage{
set_bWeightZootox_MacroPatch(bWeightZootox)
}
\arguments{
\item{bWeightZootox}{numeric}
}
\description{
Set \code{private$bWeightZootox}
}
