% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{set_conPathogen_HumanPop}
\alias{set_conPathogen_HumanPop}
\title{\code{HumanPop} Method: Set Pathogen Output Connection}
\usage{
set_conPathogen_HumanPop(conPathogen)
}
\arguments{
\item{conPathogen}{an object of class \code{\link[base]{connection}}}
}
\description{
Sets the \code{\link[base]{connection}} object that is responsible for pathogen model output
\itemize{
\item This method is bound to \code{HumanPop$set_conPathogen}
}
}
