% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{set_EIR_Human}
\alias{set_EIR_Human}
\title{Set \code{Human} EIR}
\usage{
set_EIR_Human(EIR)
}
\arguments{
\item{EIR}{numeric value}
}
\description{
Set my personal EIR value.
}
\details{
\itemize{
\item This method is bound to \code{Human$set_EIR}
}
}
