% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-ELP-Methods.R
\name{set_ELP_MacroPatch}
\alias{set_ELP_MacroPatch}
\title{Aquatic Eceology ELP: Setter for \code{\link{ELP}} in \code{\link{MacroPatch}}}
\usage{
set_ELP_MacroPatch(ELP)
}
\arguments{
\item{ELP}{new object of class \code{\link{ELP}}}
}
\description{
Set \code{\link{ELP}}
\itemize{
\item This method is bound to \code{MacroPatch$set_ELP}
}
}
