% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\name{set_homeHouseID_Human}
\alias{set_homeHouseID_Human}
\title{Human: Sets home house ID}
\usage{
set_homeHouseID_Human(homeHouseID)
}
\description{
Sets ID for home house ID
}
\details{
More details
}
