% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\name{set_homePatchID_Human}
\alias{set_homePatchID_Human}
\title{Human: Sets home patch ID}
\usage{
set_homePatchID_Human(homePatchID)
}
\description{
Sets ID for home patch ID
}
\details{
More details
}
