% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{set_humanPfMOI_Human}
\alias{set_humanPfMOI_Human}
\title{PfMOI \code{Human} Method: Set Human-stage PfMOI Object}
\usage{
set_humanPfMOI_Human(b = 0.55, c = 0.15, chemoprophylaxis = FALSE)
}
\description{
Set the \code{\link[MASHcpp]{humanPfMOI}} object in a human.
This method is bound to \code{Human$set_humanPfMOI()}
}
