% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{set_humanPfMOI_HumanPop}
\alias{set_humanPfMOI_HumanPop}
\title{PfMOI \code{HumanPop} Method: Set Human-stage PfMOI Object}
\usage{
set_humanPfMOI_HumanPop()
}
\arguments{
\item{b}{infected mosquito to human transmission efficiency}

\item{c}{infected human to mosquito transmission efficiency}
}
\description{
Set the \code{\link[MASHcpp]{humanPfMOI}} object in a human poulation.
This method is bound to \code{HumanPop$set_humanPfMOI()}
}
