% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\name{set_HumansPointer_Human}
\alias{set_HumansPointer_Human}
\title{Human: Set Human Pointer}
\usage{
set_HumansPointer_Human(HumansPointer)
}
\description{
Sets human pointer
}
\details{
More details
}
