% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Base-Class.R
\name{set_HumansPointer_Mosquito_Base}
\alias{set_HumansPointer_Mosquito_Base}
\title{Set Humans Pointer}
\usage{
set_HumansPointer_Mosquito_Base(HumansPointer)
}
\description{
Set a pointer to this \code{\link{HumanPop}} in this tile.
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_Base$get_HumansPointer}
}
}
