% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{set_kappa_MacroPatch}
\alias{set_kappa_MacroPatch}
\title{MacroPatch: Set Patch \eqn{\Kappa}}
\usage{
set_kappa_MacroPatch(kappa)
}
\arguments{
\item{kappa}{numeric}
}
\description{
\eqn{\Kappa} is the proportion of mosquitoes that would become infected after blood feeding on any human, the net infectiousness of humans.
Set \code{private$kappa}
}
