% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Base-Class.R
\name{set_MosquitoPointer_AquaPop_Base}
\alias{set_MosquitoPointer_AquaPop_Base}
\title{Set Mosquito Pointer}
\usage{
set_MosquitoPointer_AquaPop_Base(MosquitoPointer)
}
\arguments{
\item{MosquitoPointer}{a reference to \code{\link{MacroMosquito}} object}
}
\description{
Set a pointer to the mosquito population inheriting from \code{\link{Mosquito_Base}} in this tile.
}
