% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Base-Class.R
\name{set_PatchPointer_AquaPop_Base}
\alias{set_PatchPointer_AquaPop_Base}
\title{Set Patch Pointer}
\usage{
set_PatchPointer_AquaPop_Base(PatchPointer)
}
\arguments{
\item{PatchPointer}{a reference to \code{\link{MacroPatch}} object}
}
\description{
Set a pointer to the enclosing \code{\link{MacroPatch}}
}
