% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{set_resEIR_MacroPatch}
\alias{set_resEIR_MacroPatch}
\title{MacroPatch: Set fixed reservoir EIR of Patch}
\usage{
set_resEIR_MacroPatch(resEIR)
}
\arguments{
\item{reservoir}{logical}
}
\description{
Set \code{private$reservoir}
}
