% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{set_reservoir_MacroPatch}
\alias{set_reservoir_MacroPatch}
\title{MacroPatch: Set reservoir status of Patch
If the reservoir status is TRUE, will also need to set the resEIR}
\usage{
set_reservoir_MacroPatch(reservoir)
}
\arguments{
\item{reservoir}{logical}
}
\description{
Set \code{private$reservoir}
}
