% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Base-Class.R
\name{set_TilePointer_AquaPop_Base}
\alias{set_TilePointer_AquaPop_Base}
\title{Set Tile Pointer}
\usage{
set_TilePointer_AquaPop_Base(TilePointer)
}
\arguments{
\item{TilePointer}{a reference to \code{\link{MacroTile}} object}
}
\description{
Set a pointer to the enclosing \code{\link{MacroTile}}
}
