% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\name{set_TilePointer_Human}
\alias{set_TilePointer_Human}
\title{Set Tile Pointer}
\usage{
set_TilePointer_Human(TilePointer)
}
\arguments{
\item{TilePointer}{an environment}
}
\description{
Set either microsimulation \code{\link[MASHmicro]{Tile}} or macrosimulation \code{\link{Tile}} enclosing this site.
\itemize{
\item This method is bound to \code{Human$set_TilePointer}
}
}
