% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\name{set_TilePointer_MacroPatch}
\alias{set_TilePointer_MacroPatch}
\title{MacroPatch: Set Patch Pointer to Enclosing \code{\link{MacroTile}}}
\usage{
set_TilePointer_MacroPatch(TilePointer)
}
\arguments{
\item{TilePointer}{new pointer}
}
\description{
Set \code{private$TilePointer}
}
