% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Base-Class.R
\name{set_TilePointer_Mosquito_Base}
\alias{set_TilePointer_Mosquito_Base}
\title{Set Tile Pointer}
\usage{
set_TilePointer_Mosquito_Base(TilePointer)
}
\description{
Set a pointer to the enclosing \code{\link{MacroTile}}
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_Base$set_TilePointer}
}
}
