% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{set_tripDuration_HumanPop}
\alias{set_tripDuration_HumanPop}
\title{Set tripDuration}
\usage{
set_tripDuration_HumanPop(tripDuration)
}
\arguments{
\item{tripDuration}{a numeric vector of length equal to the number of humans}
}
\description{
Set the average duration of trips for each human in the population
}
