% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{set_tripFrequency_HumanPop}
\alias{set_tripFrequency_HumanPop}
\title{Set tripFrequency}
\usage{
set_tripFrequency_HumanPop(tripFrequency)
}
\arguments{
\item{tripFrequency}{a numeric vector of length equal to the number of humans}
}
\description{
Set the average frequency of trips for each human in the population
}
