% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-SimBite.R
\name{SimBitePfSI_Human}
\alias{SimBitePfSI_Human}
\title{PfSI SimBite \code{Human} Event: PfSI Simulated Bite Event}
\usage{
SimBitePfSI_Human(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of bite}

\item{PAR}{do not set to a value (default \code{NULL})}
}
\description{
Simulate a PfSI biting event. This event calls \code{\link{probeHost_PfSI}}
}
\details{
\itemize{
\item This method is bound to \code{Human$SimBitePfSI}
}
}
