% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\name{simHuman_Human}
\alias{simHuman_Human}
\title{Human: simHuman}
\usage{
simHuman_Human(tPause)
}
\arguments{
\item{tPause}{numeric; run human simulation until this point}
}
\description{
Simulate a human up until time \code{tPause}
}
