% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\name{simHumans_HumanPop}
\alias{simHumans_HumanPop}
\title{HumanPop: Simulate Humans}
\usage{
simHumans_HumanPop(tPause)
}
\arguments{
\item{tPause}{numeric (run all events in each human's \code{\link[MASHcpp]{HumanEventQ}} that occur before this time)}
}
\description{
Simulate each human's event queue.
}
