% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Simulation.R
\name{simMacro}
\alias{simMacro}
\title{Run a MACRO Simulation}
\usage{
simMacro(tMax)
}
\description{
write more stuff!!!!!!!!
}
\details{
\itemize{
\item This method is bound to \code{MacroTile$simMacro}
}
}
