% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMANS-HumanPop-Parameters.R
\name{siteAges_HumanPop}
\alias{siteAges_HumanPop}
\title{Generate Ages for \code{HumanPop}}
\usage{
siteAges_HumanPop(N)
}
\arguments{
\item{N}{size of population (may refer to household \code{\link[MASHmicro]{FeedingSite}} or an individual macrosimulation \code{\link{Patch}})}
}
\value{
numeric vector
}
\description{
Generate numeric vector of ages. This function is called in \code{\link{sitePops_HumanPop}}
}
\examples{
siteAges_HumanPop(N = 10)
}
