% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Move.R
\name{takeTrip}
\alias{takeTrip}
\title{Move \code{Human} Event: Trip Event}
\usage{
takeTrip(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of trip}

\item{PAR}{must be a list containing character \code{tDest}, the index of the site I am visiting}
}
\description{
Simulate a between patch trip.
This method is bound to \code{Human$takeTrip()}
\itemize{
\item Biting weight at my origin is decremented by \code{\link{decrement_bWeightHuman_Human}} and incremented at my destination by \code{\link{accumulate_bWeightHuman_Human}}
\item The end of this trip is queued after a duration by calling \code{\link{add2Q_returnHome}}
}
}
