% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\name{track_history_human}
\alias{track_history_human}
\title{Human: Tracks History}
\usage{
track_history_human(tEvent, event)
}
\description{
Tracks history of events for human
}
\details{
More details
}
