% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfMOI-Methods.R
\name{treatPfMOI}
\alias{treatPfMOI}
\title{PfMOI \code{Human} Event: PfMOI Treatment Event}
\usage{
treatPfMOI(tEvent, PAR)
}
\arguments{
\item{tEvent}{time of treatment}

\item{PAR}{\code{NULL}}
}
\description{
Simulate a PfMOI treatment event. If the human is infected, set susceptible and track history; also initiate period of chemoprophylaxis, see \code{\link{add2Q_endprophylaxisPfMOI}}
This method is bound to \code{Human$treatPfMOI()}
}
