% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{updateEIR_Human}
\alias{updateEIR_Human}
\title{Update my EIR}
\usage{
updateEIR_Human()
}
\description{
Based on my current location, update my daily EIR.
}
\details{
\itemize{
\item This method is bound to \code{Human$updateEIR}
}
}
