% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{updateEIR_HumanPop}
\alias{updateEIR_HumanPop}
\title{Update EIR}
\usage{
updateEIR_HumanPop()
}
\description{
Update daily EIR for all humans.
}
\details{
\itemize{
\item This method is bound to \code{HumanPop$updateEIR}
}
}
