% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{updateKappa_PfMOI_Human}
\alias{updateKappa_PfMOI_Human}
\title{PfMOI Update my Kappa}
\usage{
updateKappa_PfMOI_Human()
}
\description{
Update my personal contribution to \eqn{|Kappa} in the patch where I am currently residing.
}
\details{
\itemize{
\item This method is bound to \code{Human$updateKappa}
}
}
