% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{updateKappa_PfSI_Human}
\alias{updateKappa_PfSI_Human}
\title{PfSI Update my Kappa}
\usage{
updateKappa_PfSI_Human()
}
\description{
Update my personal contribution to \eqn{|Kappa} in the patch where I am currently residing.
}
\details{
\itemize{
\item This method is bound to \code{Human$updateKappa}
}
}
