% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\name{event_maxDeath_Human}
\alias{event_maxDeath_Human}
\title{Human: The Maximum Death Time Event Package}
\usage{
event_maxDeath_Human(tEvent = 73000, PAR = NULL, tag = "death")
}
\description{
Maximum Death Time Event Package
}
\details{
More details
}
