% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\docType{class}
\name{Human}
\alias{Human}
\title{Human Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Human
}
\description{
This is a generic human being blah blah ...
Each instance of a \code{Human} lives in a \code{\link{Human}}
}
\section{Constructor}{

\itemize{
\item myID: a character ID
\item houseID: integer ID of my current house
\item patchID: integer ID of my current patch
\item homeHouseID: integer ID of my home house
\item homePatchID: integer ID of my home patch
\item age: numeric birthday (negative; simulation begins at 0)
\item bWeight: numeric biting weight
}
}

\section{Methods}{

\itemize{
\item get_myID: see \code{\link{get_myID_Human}}
\item set_myID: see \code{\link{set_myID_Human}}
\item get_houseID: see \code{\link{get_houseID_Human}}
\item set_houseID: see \code{\link{set_houseID_Human}}
\item get_patchID: see \code{\link{get_patchID_Human}}
\item set_patchID: see \code{\link{et_patchID_Human}}
\item get_age: see \code{\link{get_age_Human}}
\item set_age: see \code{\link{set_age_Human}}
\item get_EventQ: see \code{\link{get_EventQ_Human}}
\item get_Alive: see \code{\link{get_Alive_Human}}
\item set_Alive: see \code{\link{set_Alive_Human}}
\item get_sex: see \code{\link{get_sex_Human}}
\item set_sex: see \code{\link{set_sex_Human}}
\item get_bWeight: see \code{\link{get_bWeight_Human}}
\item set_bWeight: see \code{\link{set_bWeight_Human}}
\item set_Pathogens: see \code{\link{set_Pathogens_Human}}
\item get_HumansPointer: see \code{\link{get_HumansPointer_Human}}
\item set_HumansPointer: see \code{\link{set_HumansPointer_Human}}
\item oneEvent: see \code{\link{oneEvent_Human}}
\item runEvent: see \code{\link{runEvent_Human}}
\item liveLife: see \code{\link{liveLife_Human}}
\item event_maxDeath: see \code{\link{event_maxDeath_Human}}
\item death: see \code{\link{death_Human}}
\item track_history: see \code{\link{track_history_Human}}
\item get_history: see \code{\link{get_history_Human}}
}
}

\section{Fields}{

\itemize{
\item \strong{myID}: character identifier of this human, first digits are integer ID and digits after underscore (_) are the patchID; houseID is not necessary because
location is not resolved to house level in MACRO and MICRO only needs houseID to distribute risk onto \code{\link{feedingSite}} rather than to store humans.
myID is needed to resolve location to patch level in order to move people between patches in either MICRO or MACRO.
}
}

\keyword{R6}
\keyword{class}
