% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Human-Class.R
\docType{class}
\name{Human}
\alias{Human}
\title{Human Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Human
}
\description{
This is a generic human being blah blah ...
Each instance of a \code{Human} lives in a \code{\link{HumanPop}}
}
\section{Constructor}{

\itemize{
\item item 1:
}
}

\section{Methods}{

\itemize{
\item item 1:
}
}

\section{Fields}{

\itemize{
\item \strong{myID}: character identifier of this human, first digits are integer ID and digits after underscore (_) are the patchID; houseID is not necessary because
location is not resolved to house level in MACRO and MICRO only needs houseID to distribute risk onto \code{\link{feedingSite}} rather than to store humans.
myID is needed to resolve location to patch level in order to move people between patches in either MICRO or MACRO.
}
}

\keyword{R6}
\keyword{class}
