% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Class.R
\docType{class}
\name{HumanPop}
\alias{HumanPop}
\title{MASH-MACRO HumanPop Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
HumanPop
}
\description{
generate a single human population; they may live in a MacroPatch or a MicroPatch
Each instance of a \code{Human} lives in a \code{\link{HumanPop}}
}
\section{\strong{Constructor}}{

\itemize{
\item item 1:
}
}

\section{\strong{Methods}}{

\itemize{
\item get_Human: retrieve human whose field 'myID' matches argument 'humanID' in keylist of pop field.
\item item 1:
\item item 1:
\item item 1:
\item item 1:
}
}

\section{\strong{Fields}}{

\itemize{
\item pop: a object of class \code{\link[MASHcpp]{HashMap}} that stores instantiations of \code{\link{Human}}, see help for more details on the internal structure of this type.
}
}

\keyword{R6}
\keyword{class}
