% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\docType{class}
\name{MacroPatch}
\alias{MacroPatch}
\title{MASH-MACRO MacroPatch Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MacroPatch
}
\description{
write me
}
\section{\strong{Constructor}}{

\itemize{
\item argument: i'm an argument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item travelWeight: a 'row' of a stochastic matrix describing human travel from this point (the index corresponding to self to self transition should be set to 0 as we are already conditioning on a person having decided to leave this patch)
}
}

\keyword{R6}
\keyword{class}
