% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Patch-Class.R
\docType{class}
\name{MacroPatch}
\alias{MacroPatch}
\title{MASH-MACRO MacroPatch Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MacroPatch
}
\description{
Generate a single well-mixed MacroPatch, which can run classical models in standalone mode or be linked in a network of
other patches in a \code{\link{MacroTile}}
Each instance of a \code{Human} lives in a \code{\link{HumanPop}}
}
\section{Constructor}{

\itemize{
\item item 1:
}
}

\section{Methods}{

\itemize{
\item \strong{get_Human}: retrieve human whose field 'myID' matches argument 'humanID' in keylist of pop field.
\item item 1:
\item item 1:
\item item 1:
\item item 1:
}
}

\keyword{R6}
\keyword{class}
