% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Tile-Class.R
\docType{class}
\name{MacroTile}
\alias{MacroTile}
\title{MASH-MACRO MacroTile Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MacroTile
}
\description{
write me
}
\section{\strong{Constructor}}{

\itemize{
\item nPatch: number of patches
\item AquaPar: named list of parameters for aquatic ecology model (see \code{\link{AquaPop_Emerge.Parameters}} or ... for structure)
\item PatchPar: list of parameters (length of list must be equal to number of patches and have \code{bWeightZoo} and \code{bWeightZootox} specified)
\item MosquitoPar: named list of parameters for mosquito model
\item HumanPop_PAR: list of length equal to number of humans (see \code{\link{Human}} constructor arguments for the required fields for each human)
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: i'm a field!
}
}

\keyword{R6}
\keyword{class}
