% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{queueInfectiousBites_PfSI_Human}
\alias{queueInfectiousBites_PfSI_Human}
\title{PfSI Queue my Daily Infectious Bites}
\usage{
queueInfectiousBites_PfSI_Human()
}
\description{
Based on my personal EIR value, queue infectious PfSI bites by calling \code{\link{add2Q_SimBitePfSI_Human}}
}
\details{
Each Human experiences a Poisson-distributed number of bites, with mean given by an individual's weighted EIR
\itemize{
\item This method is bound to \code{Human$queueInfectiousBites}
}
}
