% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Human-Kappa.R
\name{updateKappa_HumanPop}
\alias{updateKappa_HumanPop}
\title{Update Kappa}
\usage{
updateKappa_HumanPop()
}
\description{
Update contributions to \eqn{|Kappa} for all humans and normalize \eqn{\Kappa} for each patch by calling \code{\link{normalize_kappa_MacroPatch}}
}
\details{
\itemize{
\item This method is bound to \code{HumanPop$updateKappa}
}
}
